Move the debugger model to its own file
