+++ tests/src/service.spec.ts
@@ -6,7 +6,7 @@ import { createClientSession } from '@jupyterlab/testutils';
 
 import { PromiseDelegate } from '@phosphor/coreutils';
 
-import { Debugger } from '../../lib/debugger';
+import { DebuggerModel } from '../../lib/model';
 
 import { DebugService } from '../../lib/service';
 
@@ -59,7 +59,7 @@ describe('Debugging support', () => {
 
 describe('DebugService', () => {
   let client: IClientSession;
-  let model: Debugger.Model;
+  let model: DebuggerModel;
   let session: IDebugger.ISession;
   let service: IDebugger;
 
@@ -72,7 +72,7 @@ describe('DebugService', () => {
     await (client as ClientSession).initialize();
     await client.kernel.ready;
     session = new DebugSession({ client });
-    model = new Debugger.Model();
+    model = new DebuggerModel();
     service = new DebugService();
   });
 
@@ -115,7 +115,7 @@ describe('DebugService', () => {
 
   describe('#session', () => {
     it('should emit the sessionChanged signal when setting the session', () => {
-      let sessionChangedEvents: IDebugger.ISession[] = [];
+      const sessionChangedEvents: IDebugger.ISession[] = [];
       service.sessionChanged.connect((_, newSession) => {
         sessionChangedEvents.push(newSession);
       });
@@ -127,9 +127,9 @@ describe('DebugService', () => {
 
   describe('#model', () => {
     it('should emit the modelChanged signal when setting the model', () => {
-      let modelChangedEvents: Debugger.Model[] = [];
+      const modelChangedEvents: DebuggerModel[] = [];
       service.modelChanged.connect((_, newModel) => {
-        modelChangedEvents.push(newModel as Debugger.Model);
+        modelChangedEvents.push(newModel as DebuggerModel);
       });
       service.model = model;
       expect(modelChangedEvents.length).to.equal(1);
